refactor collector main tests to use goconvey
