+++ collector/main.go
